requirements.txt
# Core dependencies for document processing
python-docx>=1.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
chardet>=5.0.0
pdf2image>=1.16.0
Pillow>=9.0.0

# LLM Gateway dependencies
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
